rename current basic algorithm + refactor into module
